Change announcements and trending container
Change the grid layout of the announcements card to the minimum
amount need for content. This was to reduce the height of the card
to fix the stretching it would do when the width was reduced.
Update the trending section into grid layout to make it into a card
showing trending items.